backend-20: * removed mail service, moved its functionality to repository, added DTOs and test route
