+++ backend/src/core/mail/port/mail-repository.ts
@@ -1,6 +1,5 @@
-import { MailType } from '~/shared/enums/enums';
-import { MailProps } from '~/shared/types/types';
+import { MailRequestDto, MailResponseDto } from '~/shared/types/types';
 
 export interface MailRepository {
-  sendEmail(receiver: string, mailType: MailType, props: MailProps): Promise<void>;
+  sendEmail(mailRequestDto: MailRequestDto): Promise<MailResponseDto>;
 }
